requirements.txt
pyserial == 3.5
request == 2.25
